fix: router fixes
