--- apps/client/src/app/screens/login/login.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react';
-
-import Login from './login';
-
-describe('Login', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<Login />);
-    expect(baseElement).toBeTruthy();
-  });
-});
