--- apps/client/src/app/screens/login/login.tsx
@@ -1,12 +1,0 @@
-/* eslint-disable-next-line */
-export interface LoginProps {}
-
-export function Login(props: LoginProps) {
-  return (
-    <div>
-      <h1>Welcome to Login!</h1>
-    </div>
-  );
-}
-
-export default Login;
