--- apps/client/src/app/screens/register/register.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react';
-
-import Register from './register';
-
-describe('Register', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<Register />);
-    expect(baseElement).toBeTruthy();
-  });
-});
