--- apps/client/src/app/screens/register/register.tsx
@@ -1,12 +1,0 @@
-/* eslint-disable-next-line */
-export interface RegisterProps {}
-
-export function Register(props: RegisterProps) {
-  return (
-    <div>
-      <h1>Welcome to Register!</h1>
-    </div>
-  );
-}
-
-export default Register;
